Move `solana_sdk::pubkey` functions to solana-pubkey (#2980)
* move functions from solana_sdk::pubkey to solana_pubkey

* fix test

* remove serde_json from write_pubkey_file

* remove serde_json

* fix import

* put new_rand behind a separate feature

* lint

* fix imports

* fmt

* deprecate sdk::pubkey::{read_pubkey_file, write_pubkey_file} and duplicate in solana-keygen

* lint
