+++ sdk/pubkey/src/lib.rs
@@ -1114,6 +1114,12 @@ macro_rules! pubkey {
     };
 }
 
+/// New random Pubkey for tests and benchmarks.
+#[cfg(all(feature = "rand", not(target_os = "solana")))]
+pub fn new_rand() -> Pubkey {
+    Pubkey::from(rand::random::<[u8; PUBKEY_BYTES]>())
+}
+
 #[cfg(test)]
 mod tests {
     use {super::*, strum::IntoEnumIterator};
